requirements.txt
einops
gin
gin_config
librosa
local_attention
matplotlib
numpy
praat-parselmouth
pyworld
PyYAML
resampy
scikit_learn
scipy
SoundFile
torch
torchaudio
torchcrepe
tqdm
